e2e: simplify and rename procmgr test helper
Rename the shared helper package to procmgrtest and centralize dd-procmgr describe execution in the helper, with suite executors passed via a small interface to keep test call sites DRY.
