+++ test/new-e2e/internal/procmgrtest/BUILD.bazel
@@ -1,9 +1,9 @@
 load("@rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "procmgrwait",
+    name = "procmgrtest",
     srcs = ["wait.go"],
-    importpath = "github.com/DataDog/datadog-agent/test/new-e2e/internal/procmgrwait",
+    importpath = "github.com/DataDog/datadog-agent/test/new-e2e/internal/procmgrtest",
     visibility = ["//test/new-e2e:__subpackages__"],
     deps = [
         "@com_github_stretchr_testify//require",
